requirements.txt
blessings==1.7
boto3==1.9.93
botocore==1.12.93
bson==0.5.7
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
hvac==0.7.2
idna==2.8
inquirer==2.5.1
jmespath==0.9.3
pymongo==3.7.2
python-dateutil==2.8.0
readchar==2.0.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.2
google-cloud-storage==1.26.0
slackclient==2.5.0
